SNAPSHOT results web scaffold
TODOs:
- [x] import ktor server
- [x] start server with CLI
- [x] extend kodein DI in webapp
- [x] implement scaffold hello world resource
- [x] decouple resource from webapp (not single file)
- [x] test scaffold hello world resource
- [x] fix configuration saving and improve safety
- [x] consolidate kotlinx-html usage into render-html
- [x] kotlinx-html UI for hello world resource
- [x] input validation
- [x] global configuration for results webapp
- [x] refactor configuration service/repo for suspension fns and results
- [x] conditionally enable exploratory module
- [x] decompose "database" DI modules
- [x] refactor CLI command tests
      - [x] refactor config tests onto BaseConfigCommandTest
      - [x] refactor data session tests onto BaseDataSessionCommandTest
- [x] logging
- [o] integration test
      - [x] refactor executable ITs by command
      - [ ] webapp-results module ITs
